[gui] Move statistics components to separated directories
