+++ web/server/vue-cli/src/components/Statistics/Checker/CheckerStatistics.vue
@@ -1,7 +1,7 @@
 <template>
   <v-container fluid>
     <v-row>
-      <v-col cols="auto">
+      <v-col>
         <h3 class="title primary--text mb-2">
           <v-btn
             color="primary"
@@ -21,7 +21,7 @@
           </v-btn>
         </h3>
 
-        <severity-statistics-table
+        <checker-statistics-table
           :items="statistics"
           :loading="loading"
         />
@@ -31,18 +31,16 @@
 </template>
 
 <script>
-import { SeverityMixin, ToCSV } from "@/mixins";
-
-import BaseStatistics from "./BaseStatistics";
-import SeverityStatisticsTable from "./SeverityStatisticsTable";
-import { getSeverityStatistics } from "./StatisticsHelper";
+import { ReviewStatusMixin, SeverityMixin, ToCSV } from "@/mixins";
+import { BaseStatistics, getCheckerStatistics } from "@/components/Statistics";
+import CheckerStatisticsTable from "./CheckerStatisticsTable";
 
 export default {
-  name: "SeverityStatistics",
+  name: "CheckerStatistics",
   components: {
-    SeverityStatisticsTable
+    CheckerStatisticsTable
   },
-  mixins: [ BaseStatistics, SeverityMixin, ToCSV ],
+  mixins: [ BaseStatistics, ReviewStatusMixin, SeverityMixin, ToCSV ],
 
   props: {
     namespace: { type: String, required: true }
@@ -58,43 +56,38 @@ export default {
   methods: {
     downloadCSV() {
       const data = [
-        [ "Severity", "Unreviewed", "Confirmed bug",
-          "Outstanding reports (Unreviewed + Confirmed)", "False positive",
-          "Intentional", "Suppressed reports (False positive + Intentional)",
-          "All reports"
+        [
+          "Checker", "Severity", "Unreviewed",
+          "Confirmed bug", "Outstanding reports (Unreviewed + Confirmed)",
+          "False positive", "Intentional",
+          "Suppressed reports (False positive + Intentional)", "All reports"
         ],
         ...this.statistics.map(stat => {
           return [
-            this.severityFromCodeToString(stat.severity),
+            stat.checker, this.severityFromCodeToString(stat.severity),
             stat.unreviewed.count, stat.confirmed.count,
             stat.outstanding.count, stat.falsePositive.count,
-            stat.intentional.count, stat.suppressed.count, stat.reports.count
+            stat.intentional.count, stat.suppressed.count, stat.reports.count,
           ];
         })
       ];
 
-      this.toCSV(data, "codechecker_severity_statistics.csv");
+      this.toCSV(data, "codechecker_checker_statistics.csv");
     },
 
-    getStatistics: getSeverityStatistics,
+    getStatistics: getCheckerStatistics,
 
     async fetchStatistics() {
       this.loading = true;
 
       const { runIds, reportFilter, cmpData } = this.getStatisticsFilters();
       this.statistics =
-        await getSeverityStatistics(runIds, reportFilter, cmpData);
+        await getCheckerStatistics(runIds, reportFilter, cmpData);
 
-      await this.fetchDifference("severity");
+      await this.fetchDifference("component");
 
       this.loading = false;
     }
   }
 };
 </script>
-
-<style lang="scss" scoped>
-::v-deep .severity {
-  text-decoration: none;
-}
-</style>
