+++ web/server/vue-cli/src/components/Statistics/index.js
@@ -1,11 +1,32 @@
-import CheckerStatistics from "./CheckerStatistics";
-import ComponentStatistics from "./ComponentStatistics";
+import BaseStatistics from "./BaseStatistics";
+import BaseStatisticsTable from "./BaseStatisticsTable";
+import CheckerStatistics from "./Checker/CheckerStatistics";
+import CheckerStatisticsTable from "./Checker/CheckerStatisticsTable";
+import ComponentStatistics from "./Component/ComponentStatistics";
 import defaultStatisticsFilterValues from "./DefaultStatisticsFilterValues";
-import SeverityStatistics from "./SeverityStatistics";
+import ReportDiffCount from "./ReportDiffCount";
+import SeverityStatistics from "./Severity/SeverityStatistics";
+
+import {
+  getCheckerStatistics,
+  getComponentStatistics,
+  getComponents,
+  getSeverityStatistics,
+  initDiffField
+} from "./StatisticsHelper";
 
 export {
+  BaseStatistics,
+  BaseStatisticsTable,
   CheckerStatistics,
+  CheckerStatisticsTable,
   ComponentStatistics,
   defaultStatisticsFilterValues,
+  getCheckerStatistics,
+  getComponentStatistics,
+  getComponents,
+  getSeverityStatistics,
+  initDiffField,
+  ReportDiffCount,
   SeverityStatistics
 };
